Added support for template-specific item delimiter wrappers.
An item in invoice templates is delimited using a start tag --i.e. invoiceLineItem.start, and an end tag --i.e. invoiceLineItem.end. This way, an example item in the template can be replicated as many times as there are items in a given invoice.

In the LaTeX template, item delimiters don't need wrappers. They are defined as follows:

    \toprule%
    invoiceLineItem.start itemDescription.value ...
    ...
    ... & taxPercentage.value invoiceLineItem.end ...

In HTML, however, if we want our template to be valid, we need to wrap the delimiters using a tag or by commenting out them out. E.g.:

    <!--invoiceLineItem.start-->
    <tr>
      <td headers="description">itemDescription.value</td>
      ...
      <td headers="tax-ratio">taxPercentage.value</td>
    </tr>
    <!--invoiceLineItem.end-->

Clearly, if we don't use a wrapper, only the part after and before the start and the end tag (resp.) will be replicated.

 Future Development
 --------------------------------
 This mechanism could be generalized for placing arbitrary processing instructions within templates but we're good for now.