Changed the way template-specific item delimiter wrappers are implemented (cf. 3502406d32a3f0ece7ca)
The semantics are simplified as follows:

 Each template _may_ define a way to *start* and *end* a comment block by implementing getStartCommentTag() and getEndCommentTag() methods respectively.

Now, if placing 'invoiceLineItem.start' and 'invoiceLineItem.end' tags affect the template's validity --as it does in the case of HTML templates-- all we have to do is comment-out these two tags.

 This way, HTML parsers ignore the tags but OpenInvoice uses them to replicate invoice items.

 CURRENT LIMITATIONS
 ==============

 (1)

'invoiceLineItem.start' tag must immediately follow 'comment start tag'. The same is true for 'invoiceLineItem.end'.

Examples:

<!--invoiceLineItem.start--> --OK :)
<!-- invoiceLineItem.end --> --KO :(

(2)

If 'getCommentStartTag()' and 'getCommentEndTag()' methods are overridden by a template, then 'invoiceLineITem.start' and 'invoiceLineItem.end' tags MUST be wrapped with comments in the template.